zad19.3 - add NullPointerException in delete method

+++ src/main/java/com/crud/tasks/controller/TaskController.java
@@ -37,6 +37,10 @@ public class TaskController {
 
     @RequestMapping(method = RequestMethod.DELETE, value = "deleteTask")
     public void deleteTask(@RequestParam Long taskId) throws TaskNotFoundException{
+        if (taskId == null) {
+            throw new NullPointerException();
+        }
+
         try{
             service.deleteTask(taskId);
         } catch (Exception e) {
